+++ rust-inputs/input-10.rs
@@ -1,4 +1,4 @@
-fn cond(x: i32) -> i32 {
+fn cond(x: i64) -> i64 {
     if (x < 4) {
         return 0;
     } else {
@@ -6,7 +6,7 @@ fn cond(x: i32) -> i32 {
     }
 }
 
-fn main() -> i32 {
+fn main() -> i64 {
     println!("{}", cond(5));
     println!("{}", cond(3));
     return 0;
