+++ rust-inputs/input-11.rs
@@ -1,10 +1,10 @@
-fn suma(a: i32) -> i32 {
-    let mut b: i32 = 0;
+fn suma(a: i64) -> i64 {
+    let mut b: i64 = 0;
     b = 10;
     return a + b;
 }
 
-fn main() -> i32 {
+fn main() -> i64 {
     println!("{}", suma(5));
     return 0;
 }
