+++ rust-inputs/input-12.rs
@@ -1,4 +1,4 @@
-fn signo(x: i32) -> i32 {
+fn signo(x: i64) -> i64 {
     if (x < 5) {
         return 1;
     } else {
@@ -6,7 +6,7 @@ fn signo(x: i32) -> i32 {
     }
 }
 
-fn main() -> i32 {
+fn main() -> i64 {
     println!("{}", signo(3));
     println!("{}", signo(0));
     println!("{}", signo(10));
