+++ rust-inputs/input-13.rs
@@ -1,4 +1,4 @@
-fn fac(n: i32) -> i32 {
+fn fac(n: i64) -> i64 {
     if (n < 2) {
         return n;
     } else {
@@ -6,8 +6,8 @@ fn fac(n: i32) -> i32 {
     }
 }
 
-fn main() -> i32 {
-    let mut x: i32 = 0;
+fn main() -> i64 {
+    let mut x: i64 = 0;
     x = 1;
     while (x < 20) {
         println!("{}", fac(x));
