+++ rust-inputs/input-6.rs
@@ -1,8 +1,8 @@
-fn doble(x: i32) -> i32 {
+fn doble(x: i64) -> i64 {
     return x + x;
 }
 
-fn main() -> i32 {
+fn main() -> i64 {
     println!("{}", doble(21));
     return 0;
 }
