+++ struct-inputs/input-2.rs
@@ -1,6 +1,6 @@
 struct Point {
-    x: i32,
-    y: i32,
+    x: i64,
+    y: i64,
 }
 
 static ORIGIN: Point = Point { x: 0, y: 0 };
